refactor: 전체 페이지 경로 수정
